zombie villagers still need fixes
