+++ src/main/java/simplexity/villagerinfo/util/Resolvers.java
@@ -71,7 +71,7 @@ public class Resolvers {
     }
 
 
-    public TagResolver timeFormatter(Long timeDifferenceInSeconds) {
+    public TagResolver timeFormatterPast(Long timeDifferenceInSeconds) {
         Component finalTimeComponent;
         if (timeDifferenceInSeconds == null) {
             finalTimeComponent = miniMessage.deserialize(MessageInsert.NEVER_MESSAGE_FORMAT.getMessage());
@@ -100,6 +100,38 @@ public class Resolvers {
                 finalTimeComponent = finalTimeComponent.append(miniMessage.deserialize(MessageInsert.JUST_NOW_FORMAT.getMessage()));
             } else {
                 finalTimeComponent = finalTimeComponent.append(miniMessage.deserialize(MessageInsert.AGO_MESSAGE_FORMAT.getMessage()));
+            }
+        }
+        return TagResolver.resolver(Placeholder.component("time", finalTimeComponent));
+    }
+
+    public TagResolver timeFormatter(Long timeDifferenceInSeconds) {
+        Component finalTimeComponent;
+        if (timeDifferenceInSeconds == null) {
+            finalTimeComponent = miniMessage.deserialize(MessageInsert.NEVER_MESSAGE_FORMAT.getMessage());
+        } else {
+            long s = timeDifferenceInSeconds % 60;
+            long m = (timeDifferenceInSeconds / 60) % 60;
+            long h = (timeDifferenceInSeconds / (60 * 60)) % 24;
+            finalTimeComponent = Component.empty();
+            boolean componentEmpty = true;
+            if (h > 0) {
+                componentEmpty = false;
+                String hours = String.valueOf(h);
+                finalTimeComponent = finalTimeComponent.append(miniMessage.deserialize(MessageInsert.HOUR_MESSAGE_FORMAT.getMessage(), Placeholder.parsed("value", hours)));
+            }
+            if (m > 0) {
+                componentEmpty = false;
+                String minutes = String.valueOf(m);
+                finalTimeComponent = finalTimeComponent.append(miniMessage.deserialize(MessageInsert.MINUTE_MESSAGE_FORMAT.getMessage(), Placeholder.parsed("value", minutes)));
+            }
+            if (s > 0) {
+                componentEmpty = false;
+                String seconds = String.valueOf(s);
+                finalTimeComponent = finalTimeComponent.append(miniMessage.deserialize(MessageInsert.SECOND_MESSAGE_FORMAT.getMessage(), Placeholder.parsed("value", seconds)));
+            }
+            if (componentEmpty) {
+                finalTimeComponent = finalTimeComponent.append(miniMessage.deserialize(MessageInsert.JUST_NOW_FORMAT.getMessage()));
             }
         }
         return TagResolver.resolver(Placeholder.component("time", finalTimeComponent));
